Fixed visualization error with score display across worlds

+++ src/gen.cpp
@@ -238,7 +238,7 @@ void generateWorld(Engine &ctx) {
         ctx.get<Attributes>(agent) = {1 - i*DEFENDER_SLOWDOWN, 0.f, 0.f, 6.5f, ctx.get<Position>(agent).position};
 
         // Use actual hoop entity IDs from gameState
-        uint32_t defending_hoop_id = (i % 2 == 0) ? gameState.team0Hoop : gameState.team1Hoop;
+        uint32_t defending_hoop_id = (i % 2 == 0) ? (uint32_t)gameState.team0Hoop : (uint32_t)gameState.team1Hoop;
         ctx.get<Team>(agent) = Team{i % 2, team_colors[i % 2], defending_hoop_id};
     };
 
@@ -346,7 +346,7 @@ void resetWorld(Engine &ctx) {
 
         ctx.get<Attributes>(agent) = {1.f - agent_i*DEFENDER_SLOWDOWN, 0.f, 0.f, 6.5f, pos.position};
 
-        uint32_t defending_hoop_id = (agent_i % 2 == 0) ? gameState.team0Hoop : gameState.team1Hoop;
+        uint32_t defending_hoop_id = (agent_i % 2 == 0) ? (uint32_t)gameState.team0Hoop : (uint32_t)gameState.team1Hoop;
         ctx.get<Team>(agent) = Team{agent_i % 2, team_colors[agent_i % 2], defending_hoop_id};
 
         agent_i++;
